Add Eventmanager and Update Attack Push Movement
Removed OnUpdate, OnHitConfirm, OnHitConfirmPauseEnd, OnPush, OnAnimationComplete and OnCoolDownComplete form the Attack script and LocalPlayerManager. Added a Standlone EventManager script to handles events. Updated all the corresponding subscribing methods in the MovementController, AttackController, TeamController, UerInterfaceController, and CameraControler to match the events Signature. Added hitbox and hurtbox to OnHitConfirmPauseend. Removed the direction parameter form OnhitConfirmPauseEnd method and event and instead pass the hitbox and hurtbox in the MovementController. Added the direction to OnPush event and method and inverse transform it  to the forceDirection and add it to the relativevelocity in the MoveContorller. Also added a isPush variable to make sure the character doesn't face the direction their being push on line 89 in the MoveController.